add logs saving endpoint & logs getting endpoint

+++ src/models/log.rs
@@ -1,9 +1,9 @@
 use clickhouse::Row;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-#[derive(Deserialize, Debug, Row)]
+#[derive(Serialize, Deserialize, Debug, Row)]
 pub struct Log {
-    pub timestamp: i64,
+    pub timestamp: u32,
     pub level: String,
     pub service: String,
     pub message: String,
